requirements.txt
flask
flask_cors
line-bot-sdk
gunicorn
google-cloud-firestore
google-cloud-logging
google-cloud-storage
grpcio
pandas
Pillow
tensorflow-cpu
wheel